// src/utils/enums/productInCart.ts
export const enum EProductInCart {
  CLASS_PRODUCTS_ITEM = 'products__item',
  CLASS_ITEM = 'item',
  CLASS_ITEM_NUMBER = 'item__number',
  CLASS_ITEM_CONTENT = 'item__content',
  CLASS_ITEM_IMG = 'item__img',
  CLASS_ITEM_INFO = 'item__info',
  CLASS_ITEM_TITLE = 'item__title',
  CLASS_ITEM_DESCRIPTION = 'item__description',
  CLASS_ITEM_META = 'item__meta',
  CLASS_ITEM_RATING = 'item__rating',
  CLASS_ITEM_DISCOUNT = 'item__discount',
  CLASS_ITEM_STOCK = 'item__stock',
  CLASS_ITEM_SUMMARY = 'item__summary',
  CLASS_ITEM_QUANTITY = 'item__quantity',
  CLASS_ITEM_BTN_LEFT = 'item__btn_left',
  CLASS_ITEM_QUANTITY_NUMBER = 'item__quantity-number',
  CLASS_ITEM_BTN_RIGHT = 'item__btn_right',
  CLASS_ITEM_SUM = 'item__sum',
  CLASS_ITEM_CATEGORY = 'item__category',
}